platform specific refactorings
