+++ Solution/WindowsAndroid/Android/V3DEngine/V3DEngine.vcxproj
@@ -37,7 +37,6 @@
   <ItemGroup>
     <ClInclude Include="..\..\..\..\Source\Android\V3DEngine\android_native_app_glue.h" />
     <ClInclude Include="..\..\..\..\Source\Android\V3DEngine\android_native_app_glue.hpp" />
-    <ClInclude Include="..\..\..\..\Source\Android\V3DEngine\V3DCore\V3DEnvironment.h" />
     <ClInclude Include="..\..\..\..\Source\Android\V3DEngine\V3DEngineLibrary.h" />
     <ClInclude Include="..\..\..\..\Source\Android\V3DEngine\V3DIO\V3DOstream.h" />
     <ClInclude Include="..\..\..\..\Source\Android\V3DEngine\V3DNetwork\V3DNetworkSystem.h" />
