+++ Solution/WindowsAndroid/Android/V3DEngine/V3DEngine.vcxproj.filters
@@ -21,9 +21,6 @@
     <ClInclude Include="..\..\..\..\Source\Android\V3DEngine\V3DIO\V3DOstream.h">
       <Filter>Source\V3DIO</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\..\..\Source\Android\V3DEngine\V3DCore\V3DEnvironment.h">
-      <Filter>Source\V3DCore</Filter>
-    </ClInclude>
     <ClInclude Include="..\..\..\..\Source\Android\V3DEngine\android_native_app_glue.h">
       <Filter>Source</Filter>
     </ClInclude>
