+++ Source/Windows/V3DEngine/V3DCore/V3DEnvironment.cpp
@@ -4,7 +4,7 @@ Copyright (c) 2020 by 4D Illusions. All rights reserved.
 Released under the terms of the GNU General Public License version 3 or later.
 */
 
-#include "V3DEnvironment.h"
+#include "V3DEngine/V3DCore/V3DEnvironment.h"
 #include "V3DEngine/V3DCore/V3DAssetPathType.h"
 
 #include <Windows.h>
@@ -20,9 +20,6 @@ namespace V3D::V3DEngine::V3DCore
 				return "Windows x86";
 			#endif
 		#endif
-
-		// ReSharper disable once CppUnreachableCode
-		return "";
 	}
 
 	int V3DEnvironment::GetWindowWidth() const
@@ -63,7 +60,9 @@ namespace V3D::V3DEngine::V3DCore
 
 			case V3DAssetPathType::Internal: return "";
 
-			default: return "default";
+			//default: return "default";
 		}
+
+		return nullptr;
 	}
 }
